+++ example/lib/barcode_scanner_controller.dart
@@ -14,10 +14,11 @@ class _BarcodeScannerWithControllerState
     extends State<BarcodeScannerWithController>
     with SingleTickerProviderStateMixin {
 
-  Barcode? barcode;
+  BarcodeCapture? barcode;
 
   MobileScannerController controller = MobileScannerController(
     torchEnabled: true,
+    detectionSpeed: DetectionSpeed.unrestricted
     // formats: [BarcodeFormat.qrCode]
     // facing: CameraFacing.front,
   );
@@ -100,7 +101,7 @@ class _BarcodeScannerWithControllerState
                           height: 50,
                           child: FittedBox(
                             child: Text(
-                              barcode?.rawValue ?? 'Scan something!',
+                              barcode?.barcodes.first.rawValue ?? 'Scan something!',
                               overflow: TextOverflow.fade,
                               style: Theme.of(context)
                                   .textTheme