+++ lib/src/barcode_capture.dart
@@ -1,14 +1,14 @@
 import 'dart:typed_data';
 
-import 'package:mobile_scanner/src/objects/barcode.dart';
+import 'package:mobile_scanner/src/barcode.dart';
 
 class BarcodeCapture {
   List<Barcode> barcodes;
-  Uint8List image;
+  Uint8List? image;
 
   BarcodeCapture({
     required this.barcodes,
-    required this.image
+    this.image,
 });
 
 }