+++ lib/src/enums/detection_speed.dart
@@ -1,0 +1,10 @@
+/// The facing of a camera.
+enum DetectionSpeed {
+  noDuplicates,
+
+  /// Front facing camera.
+  normal,
+
+  /// Back facing camera.
+  unrestricted,
+}