+++ lib/src/enums/ratio.dart
@@ -1,0 +1,1 @@
+enum Ratio { ratio_4_3, ratio_16_9 }