+++ lib/src/mobile_scanner.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:mobile_scanner/mobile_scanner.dart';
-import 'package:mobile_scanner/src/objects/barcode_capture.dart';
+import 'package:mobile_scanner/src/mobile_scanner_arguments.dart';
+import 'package:mobile_scanner/src/mobile_scanner_controller.dart';
+import 'package:mobile_scanner/src/barcode_capture.dart';
 
 /// A widget showing a live camera preview.
 class MobileScanner extends StatefulWidget {
@@ -11,32 +12,24 @@ class MobileScanner extends StatefulWidget {
   /// Function that gets called when a Barcode is detected.
   ///
   /// [barcode] The barcode object with all information about the scanned code.
-  /// [arguments] Information about the state of the MobileScanner widget
-  final Function(Barcode barcode, MobileScannerArguments? args) onDetect;
-
-  /// Function that gets called when a barcode or multiple barcodes are detected.
-  final Function(BarcodeCapture barcodeCapture, MobileScannerArguments? args)? onDetectList;
-
-  /// TODO: Function that gets called when the Widget is initialized. Can be usefull
-  /// to check wether the device has a torch(flash) or not.
-  ///
-  /// [arguments] Information about the state of the MobileScanner widget
-  // final Function(MobileScannerArguments args)? onInitialize;
+  /// [startArguments] Information about the state of the MobileScanner widget
+  final Function(
+          BarcodeCapture capture, MobileScannerArguments? arguments)
+      onDetect;
 
   /// Handles how the widget should fit the screen.
   final BoxFit fit;
 
-  /// Set to false if you don't want duplicate scans.
-  final bool allowDuplicates;
+  /// Whether to automatically resume the camera when the application is resumed
+  final bool autoResume;
 
   /// Create a [MobileScanner] with a [controller], the [controller] must has been initialized.
   const MobileScanner({
     super.key,
     required this.onDetect,
-    this.onDetectList,
     this.controller,
+    this.autoResume = true,
     this.fit = BoxFit.cover,
-    this.allowDuplicates = false,
   });
 
   @override
@@ -59,7 +52,7 @@ class _MobileScannerState extends State<MobileScanner>
   void didChangeAppLifecycleState(AppLifecycleState state) {
     switch (state) {
       case AppLifecycleState.resumed:
-        if (!controller.isStarting && controller.autoResume) controller.start();
+        if (!controller.isStarting && widget.autoResume) controller.start();
         break;
       case AppLifecycleState.inactive:
       case AppLifecycleState.paused:
@@ -69,32 +62,18 @@ class _MobileScannerState extends State<MobileScanner>
     }
   }
 
-  String? lastScanned;
-
   @override
   Widget build(BuildContext context) {
     return ValueListenableBuilder(
-      valueListenable: controller.arguments,
+      valueListenable: controller.startArguments,
       builder: (context, value, child) {
         value = value as MobileScannerArguments?;
         if (value == null) {
           return const ColoredBox(color: Colors.black);
         } else {
           controller.barcodes.listen((barcode) {
-            if (!widget.allowDuplicates) {
-              if (lastScanned != barcode.rawValue) {
-                lastScanned = barcode.rawValue;
-                widget.onDetect(barcode, value! as MobileScannerArguments);
-              }
-            } else {
-              widget.onDetect(barcode, value! as MobileScannerArguments);
-            }
+            widget.onDetect(barcode, value! as MobileScannerArguments);
           });
-          if (widget.onDetectList != null) {
-            controller.barcodesList.listen((barcode) {
-              widget.onDetectList!(barcode, value! as MobileScannerArguments);
-            });
-          }
           return ClipRect(
             child: SizedBox(
               width: MediaQuery.of(context).size.width,